Revert "Address Coverity CID 161765: Integer handling issues (SIGN_EXTENSION)."
This reverts commit 0ecdf6820643cb9482c97ded9e0dc34e91f03668.

Bad patch, now these warnings are back:

../../libparse/binio.c: In function 'get_lsb_ulong':
../../libparse/binio.c:43:10: warning: conversion to 'long unsigned int' from 'int' may change the sign of the result [-Wsign-conversion]
   retval |= *((*bufpp)++) << 8;
          ^~
../../libparse/binio.c:44:10: warning: conversion to 'long unsigned int' from 'int' may change the sign of the result [-Wsign-conversion]
   retval |= *((*bufpp)++) << 16;
          ^~
../../libparse/binio.c:45:10: warning: conversion to 'long unsigned int' from 'int' may change the sign of the result [-Wsign-conversion]
   retval |= *((*bufpp)++) << 24;
          ^~

+++ libparse/binio.c
@@ -37,14 +37,14 @@ get_lsb_ulong(
 	unsigned char **bufpp
 	)
 {
-  unsigned long retval;
+  long retval;
 
   retval  = *((*bufpp)++);
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++) << 16;
   retval |= *((*bufpp)++) << 24;
 
-  return retval;
+  return (unsigned long)retval;
 }
 
 void
